Refactor admin settings module structure
